feat(v-app): moving functionality
moved `v-app` mixins directly onto the $vuetify object
