+++ src/components/VApp/VApp.js
@@ -1,19 +1,10 @@
 import '../../stylus/components/_app.styl'
 
-// Component level mixins
-import AppTheme from './mixins/app-theme'
-import AppBreakpoint from './mixins/app-breakpoint'
-
 // Directives
 import Resize from '../../directives/resize'
 
 export default {
   name: 'v-app',
-
-  mixins: [
-    AppBreakpoint,
-    AppTheme
-  ],
 
   directives: {
     Resize
@@ -35,10 +26,12 @@ export default {
     }
   },
 
+  // TODO: Deprecate
   mounted () {
     this.$vuetify.dark = this.dark
   },
 
+  // TODO: Deprecate
   watch: {
     dark () {
       this.$vuetify.dark = this.dark
@@ -50,11 +43,7 @@ export default {
       staticClass: 'application',
       'class': this.classes,
       attrs: { 'data-app': true },
-      domProps: { id: this.id },
-      directives: [{
-        name: 'resize',
-        value: this.onResize
-      }]
+      domProps: { id: this.id }
     }
 
     const wrapper = h('div', { staticClass: 'application--wrap' }, this.$slots.default)
