+++ src/components/Vuetify/mixins/methods.js
@@ -1,0 +1,7 @@
+import goTo from '../util/goTo'
+
+export default {
+  methods: {
+    goTo
+  }
+}
