+++ test/unit/components/VApp/VApp.spec.js
@@ -40,13 +40,4 @@ test('VApp.js', ({ mount }) => {
     })
     expect(wrapper.vm.$vuetify.dark).toBe(false)
   })
-
-  it('should watch theme', async () => {
-    const wrapper = mount(VApp)
-
-    expect(wrapper.vm.style).toMatchSnapshot()
-    wrapper.vm.$vuetify.theme.primary = '#000'
-    await wrapper.vm.$nextTick()
-    expect(wrapper.vm.style).toMatchSnapshot()
-  })
 })
